% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarywidget.R
\name{summarywidget}
\alias{summarywidget}
\title{Show a single summary statistic in a widget}
\usage{
summarywidget(
  data,
  statistic = c("count", "sum", "mean"),
  column = NULL,
  selection = NULL,
  locales = NULL,
  options = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Data to summarize, normally an instance of \link[crosstalk:SharedData]{crosstalk::SharedData}.}

\item{statistic}{The statistic to compute.}

\item{column}{For \code{sum} and \code{mean} statistics, the column of \code{data} to summarize.
Not used for \code{count} statistic.}

\item{selection}{Expression to select a fixed subset of \code{data}. May be
a logical vector or a one-sided formula that evaluates to a logical vector.
If used, the \code{key} given to \link[crosstalk:SharedData]{crosstalk::SharedData} must be a fixed column (not row numbers).}

\item{locales}{A locales string accepted by
\url{https://www.w3schools.com/jsref/jsref_tolocalestring_number.asp}.}

\item{options}{A list of options accepted by
\url{https://www.w3schools.com/jsref/jsref_tolocalestring_number.asp}.
Ignored if \code{locales} argument is not supplied.}

\item{width}{Currently not used.}

\item{height}{Currently not used.}

\item{elementId}{Currently not used.}
}
\description{
A \code{summarywidget} displays a single statistic derived from a linked table.
Its primary use is with the \code{crosstalk} package. Used with \code{crosstalk},
a \code{summarywidget} displays a value which updates as the data selection
changes.
}
