% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarywidget.R
\name{summarywidget}
\alias{summarywidget}
\title{Widget that shows a single summary statistic}
\usage{
summarywidget(data, selection, statistic = c("count", "sum", "mean"),
  column = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Data to summarize, normally an instance \link[crosstalk:SharedData]{crosstalk::SharedData}.}

\item{selection}{Expression to select a fixed subset of \code{data}. May be
a boolean vector or a one-sided formula.}

\item{statistic}{The statistic to compute.}

\item{column}{For \code{sum} and \code{mean} statistics, the column of \code{data} to summarize.
Not used for \code{count} statistic.}
}
\description{
\code{summarywidget} displays a single statistic derived from a linked table.
It's primary use is with the \code{crosstalk} package. Used with \code{crosstalk},
a \code{summarywidget} displays a value which updates as the data selection
changes.
}
